+++ Src/CSharpier.VSCode/src/NullCSharpierProcess.ts
@@ -1,9 +1,5 @@
-import { Disposable } from "vscode";
 import { Logger } from "./Logger";
-
-export interface ICSharpierProcess extends Disposable {
-    formatFile(content: string, filePath: string): Promise<string>;
-}
+import { ICSharpierProcess } from "./ICSharpierProcess";
 
 export class NullCSharpierProcess implements ICSharpierProcess {
     static instance: NullCSharpierProcess;
@@ -17,10 +13,20 @@ export class NullCSharpierProcess implements ICSharpierProcess {
         this.logger = logger;
     }
 
+    getProcessFailedToStart(): boolean {
+        return false;
+    }
+
     formatFile(content: string, filePath: string): Promise<string> {
-        this.logger.debug("Skipping formatting because this is a NullCSharpierProcess. This generally indicates there was a problem starting the CSharpier process");
+        this.logger.debug(
+            "Skipping formatting because this is a NullCSharpierProcess. This generally indicates there was a problem starting the CSharpier process",
+        );
         return Promise.resolve("");
     }
 
     dispose() {}
+
+    getVersion(): string {
+        return "NULL";
+    }
 }